#1. Rename "Response" classes to "Direct"
