+++ src/types/direct/clean/CleanBirthdateDirect.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace nsusoft\dadata\types\direct\clean;
+
+use nsusoft\dadata\types\interfaces\clean\CleanBirthdateInterface;
+
+class CleanBirthdateDirect extends CleanDirect implements CleanBirthdateInterface
+{
+
+}
