+++ src/types/direct/clean/CleanEmailDirect.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace nsusoft\dadata\types\direct\clean;
+
+use nsusoft\dadata\types\interfaces\clean\CleanEmailInterface;
+
+class CleanEmailDirect extends CleanDirect implements CleanEmailInterface
+{
+
+}
