+++ src/types/direct/clean/CleanNameDirect.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace nsusoft\dadata\types\direct\clean;
+
+use nsusoft\dadata\types\interfaces\clean\CleanNameInterface;
+
+class CleanNameDirect extends CleanDirect implements CleanNameInterface
+{
+
+}
