+++ src/types/direct/clean/CleanPassportDirect.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace nsusoft\dadata\types\direct\clean;
+
+use nsusoft\dadata\types\interfaces\clean\CleanPassportInterface;
+
+class CleanPassportDirect extends CleanDirect implements CleanPassportInterface
+{
+
+}
