+++ src/types/direct/clean/CleanPhoneDirect.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace nsusoft\dadata\types\direct\clean;
+
+use nsusoft\dadata\types\interfaces\clean\CleanPhoneInterface;
+
+class CleanPhoneDirect extends CleanDirect implements CleanPhoneInterface
+{
+
+}
