+++ src/types/direct/clean/CleanVehicleDirect.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace nsusoft\dadata\types\direct\clean;
+
+use nsusoft\dadata\types\interfaces\clean\CleanVehicleInterface;
+
+class CleanVehicleDirect extends CleanDirect implements CleanVehicleInterface
+{
+
+}
