--- src/types/responses/clean/CleanBirthdateResponse.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace nsusoft\dadata\types\responses\clean;
-
-use nsusoft\dadata\types\interfaces\clean\CleanBirthdateInterface;
-
-class CleanBirthdateResponse extends CleanResponse implements CleanBirthdateInterface
-{
-
-}
