--- src/types/responses/clean/CleanEmailResponse.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace nsusoft\dadata\types\responses\clean;
-
-use nsusoft\dadata\types\interfaces\clean\CleanEmailInterface;
-
-class CleanEmailResponse extends CleanResponse implements CleanEmailInterface
-{
-
-}
