--- src/types/responses/clean/CleanNameResponse.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace nsusoft\dadata\types\responses\clean;
-
-use nsusoft\dadata\types\interfaces\clean\CleanNameInterface;
-
-class CleanNameResponse extends CleanResponse implements CleanNameInterface
-{
-
-}
