--- src/types/responses/clean/CleanPassportResponse.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace nsusoft\dadata\types\responses\clean;
-
-use nsusoft\dadata\types\interfaces\clean\CleanPassportInterface;
-
-class CleanPassportResponse extends CleanResponse implements CleanPassportInterface
-{
-
-}
