--- src/types/responses/clean/CleanPhoneResponse.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace nsusoft\dadata\types\responses\clean;
-
-use nsusoft\dadata\types\interfaces\clean\CleanPhoneInterface;
-
-class CleanPhoneResponse extends CleanResponse implements CleanPhoneInterface
-{
-
-}
