--- src/types/responses/clean/CleanVehicleResponse.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace nsusoft\dadata\types\responses\clean;
-
-use nsusoft\dadata\types\interfaces\clean\CleanVehicleInterface;
-
-class CleanVehicleResponse extends CleanResponse implements CleanVehicleInterface
-{
-
-}
